fix(server): make CLI entrypoints consistent
