+++ packages/api-server/src/index.ts
@@ -1,4 +1,5 @@
 #!/usr/bin/env node
+
 import { hideBin } from 'yargs/helpers'
 import yargs from 'yargs/yargs'
 
@@ -13,29 +14,38 @@ import {
 
 export * from './types'
 
-const positionalArgs = yargs(hideBin(process.argv)).parseSync()._
-
-// "bin": {
-//   "rw-api-server-watch": "./dist/watch.js",
-//   "rw-log-formatter": "./dist/logFormatter/bin.js",
-//   "rw-server": "./dist/index.js"
-// },
-
 if (require.main === module) {
-  if (positionalArgs.includes('api') && !positionalArgs.includes('web')) {
-    apiServerHandler(
-      yargs(hideBin(process.argv)).options(apiCliOptions).parseSync()
+  yargs(hideBin(process.argv))
+    .scriptName('rw-server')
+    .usage('usage: $0 <side>')
+    .strict()
+
+    .command(
+      '$0',
+      'Run both api and web servers',
+      // @ts-expect-error just passing yargs though
+      (yargs) => {
+        yargs.options(commonOptions)
+      },
+      bothServerHandler
     )
-  } else if (
-    positionalArgs.includes('web') &&
-    !positionalArgs.includes('api')
-  ) {
-    webServerHandler(
-      yargs(hideBin(process.argv)).options(webCliOptions).parseSync()
+    .command(
+      'api',
+      'Start server for serving only the api',
+      // @ts-expect-error just passing yargs though
+      (yargs) => {
+        yargs.options(apiCliOptions)
+      },
+      apiServerHandler
     )
-  } else {
-    bothServerHandler(
-      yargs(hideBin(process.argv)).options(commonOptions).parseSync()
+    .command(
+      'web',
+      'Start server for serving only the web side',
+      // @ts-expect-error just passing yargs though
+      (yargs) => {
+        yargs.options(webCliOptions)
+      },
+      webServerHandler
     )
-  }
+    .parse()
 }
