+++ tasks/server-tests/jest.config.js
@@ -1,7 +1,9 @@
 /** @type {import('jest').Config} */
 const config = {
   rootDir: '.',
+  testMatch: ['<rootDir>/*.test.mjs'],
   testTimeout: 5_000 * 2,
+  transform: {},
 }
 
 module.exports = config
